move `PersistentContainer` to its own production file
